Moved fido2 models to vault in libs
