+++ libs/common/src/vault/api/fido2-key.api.ts
@@ -1,4 +1,4 @@
-import { BaseResponse } from "../../../models/response/base.response";
+import { BaseResponse } from "../../models/response/base.response";
 
 export class Fido2KeyApi extends BaseResponse {
   nonDiscoverableId: string;
