+++ libs/common/src/vault/models/data/login.data.ts
@@ -1,6 +1,6 @@
-import { Fido2KeyData } from "../../../fido2/models/data/fido2-key.data";
 import { LoginApi } from "../../../models/api/login.api";
 
+import { Fido2KeyData } from "./fido2-key.data";
 import { LoginUriData } from "./login-uri.data";
 
 export class LoginData {
