+++ libs/common/src/vault/models/domain/fido2-key.ts
@@ -1,10 +1,10 @@
 import { Jsonify } from "type-fest";
 
+import { Fido2KeyView } from "../../../fido2/models/view/fido2-key.view";
 import Domain from "../../../models/domain/domain-base";
 import { EncString } from "../../../models/domain/enc-string";
 import { SymmetricCryptoKey } from "../../../models/domain/symmetric-crypto-key";
 import { Fido2KeyData } from "../data/fido2-key.data";
-import { Fido2KeyView } from "../view/fido2-key.view";
 
 export class Fido2Key extends Domain {
   nonDiscoverableId: EncString | null = null;
