+++ libs/common/src/vault/models/domain/login.ts
@@ -1,12 +1,12 @@
 import { Jsonify } from "type-fest";
 
-import { Fido2Key } from "../../../fido2/models/domain/fido2-key";
 import Domain from "../../../models/domain/domain-base";
 import { EncString } from "../../../models/domain/enc-string";
 import { SymmetricCryptoKey } from "../../../models/domain/symmetric-crypto-key";
 import { LoginData } from "../data/login.data";
 import { LoginView } from "../view/login.view";
 
+import { Fido2Key } from "./fido2-key";
 import { LoginUri } from "./login-uri";
 
 export class Login extends Domain {
