+++ libs/common/src/vault/models/request/cipher.request.ts
@@ -1,10 +1,10 @@
-import { Fido2KeyApi } from "../../../fido2/models/api/fido2-key.api";
 import { CardApi } from "../../../models/api/card.api";
 import { FieldApi } from "../../../models/api/field.api";
 import { IdentityApi } from "../../../models/api/identity.api";
 import { LoginUriApi } from "../../../models/api/login-uri.api";
 import { LoginApi } from "../../../models/api/login.api";
 import { SecureNoteApi } from "../../../models/api/secure-note.api";
+import { Fido2KeyApi } from "../../api/fido2-key.api";
 import { CipherRepromptType } from "../../enums/cipher-reprompt-type";
 import { CipherType } from "../../enums/cipher-type";
 import { Cipher } from "../domain/cipher";
