+++ libs/common/src/vault/models/response/cipher.response.ts
@@ -1,10 +1,10 @@
-import { Fido2KeyApi } from "../../../fido2/models/api/fido2-key.api";
 import { CardApi } from "../../../models/api/card.api";
 import { FieldApi } from "../../../models/api/field.api";
 import { IdentityApi } from "../../../models/api/identity.api";
 import { LoginApi } from "../../../models/api/login.api";
 import { SecureNoteApi } from "../../../models/api/secure-note.api";
 import { BaseResponse } from "../../../models/response/base.response";
+import { Fido2KeyApi } from "../../api/fido2-key.api";
 import { CipherRepromptType } from "../../enums/cipher-reprompt-type";
 
 import { AttachmentResponse } from "./attachment.response";
