+++ libs/common/src/vault/services/cipher.service.ts
@@ -6,7 +6,6 @@ import { SearchService } from "../../abstractions/search.service";
 import { SettingsService } from "../../abstractions/settings.service";
 import { StateService } from "../../abstractions/state.service";
 import { FieldType, UriMatchType } from "../../enums";
-import { Fido2Key } from "../../fido2/models/domain/fido2-key";
 import { sequentialize } from "../../misc/sequentialize";
 import { Utils } from "../../misc/utils";
 import Domain from "../../models/domain/domain-base";
@@ -22,6 +21,7 @@ import { CipherData } from "../models/data/cipher.data";
 import { Attachment } from "../models/domain/attachment";
 import { Card } from "../models/domain/card";
 import { Cipher } from "../models/domain/cipher";
+import { Fido2Key } from "../models/domain/fido2-key";
 import { Field } from "../models/domain/field";
 import { Identity } from "../models/domain/identity";
 import { Login } from "../models/domain/login";
